re PR fortran/32770 ([Meta-bug] -fdefault-integer-8 issues)
2008-05-04  Thomas Koenig  <tkoenig@gcc.gnu.org>

	PR fortran/32770
	* gfortran.dg/any_all_1.f90:  Adjust kinds to make test
	pass with -fdefault-integer-8.
	* gfortran.dg/maxloc_bounds_4.f90:  Likewise.
	* gfortran.dg/maxloc_bounds_5.f90:  Likewise.
	* gfortran.dg/maxloc_bounds_7.f90:  Likewise.

From-SVN: r134936

+++ gcc/testsuite/gfortran.dg/any_all_1.f90
@@ -4,7 +4,7 @@
 program main
   real, dimension(2,2) :: a
   logical(kind=4), dimension(2) :: b
-  integer, dimension(2) :: i
+  integer(kind=4), dimension(2) :: i
   equivalence (b,i)
   data a /1.0, 2.0, -0.1, -0.2 /
